caesar/cs50Final/requirements.txt
cs50~=9.3.4
Flask~=3.0.3
Flask-Session
pytz~=2024.1
requests~=2.32.3

pip~=23.2.1
wheel~=0.43.0
Jinja2~=3.1.4
termcolor~=2.4.0
typing_extensions~=4.12.0
MarkupSafe~=2.1.5
Werkzeug~=3.0.3
click~=8.1.7
blinker~=1.8.2
itsdangerous~=2.2.0
certifi~=2024.7.4
msgspec~=0.18.6
urllib3~=2.2.2
cachelib~=0.13.0
greenlet~=3.0.3
idna~=3.7
sqlparse~=0.5.1
SQLAlchemy~=2.0.31
python-dotenv~=1.0.1